Update rental_shop.py to revise a RentalShop field
Originally RentalShop objects were to connect to an SQLite database
consisting of all rentable cars, details of the different costs for
renting a car, and a listing of all the current rentals.

But for the sake of simplicity, and not going too far beyond the
material covered in the university module the program was created
for (as coursework), CSV files will be used instead to implement
the database.