Add database initialisation to rental_shop.py
Functionality was added to the rental_shop.py file for setting up
default databases for RentalShop objects that do not already have
one. Three files will be created:

* cars.csv: stores the cars that exist in the rental shop. By
default 10 records will be created.

* car_rentals.csv: stores records for all the current cars being
rented by customers, along with the charged rate and time period.

* car_types.csv: stores the different car types that exist along
with their rental pricing information.

They are stored in such a way that enables them to be imported to
and exported from a pandas DataFrame.

The init() method of main.py was also modified (still unfinished)
to create a new rental shop as a form of testing. It should cause
a new directory "./.databases/shop1" to be created consisting of
the files above.