Update main.py to finish implementing init()
The init() function (now called init_objects() for better clarity)
has now been properly implemented. It prompts the user to enter
a numerical identifier (which is validated), which is then used
to create a Customer object.

The function now returns a tuple containing the Customer object
followed by the RentalShop object created earlier, so that it can
be accessed by the main() function.