Update source files to properly implement inquire()
The inqire() function of the main.py function has now been
properly implemented. Most of the implementation is done in
rental_shop.py.

In particular, the display_shop() method of the
RentalShop class imports the datasets (using a newly defined
auxiliary method called _get_data()), then transforms the 'cars'
table to produce the stock of each car type with available cars.
It displays this to the user followed by the 'car_types' table
giving the pricing information.

The pd.DataFame.to_string() method saved a lot of work with having
to output the sock and prices in a neat format.