Update source files to implement car renting.
The rent() module of main.py has now been properly implemented
(though not yet fully). A user can now rent a car for a specific
number of days, and it will be recorded on the rental shop's
database, such that the rental is still maintained after the
program terminates, and so will be remembered by the system when
the program is run again.

The _get_data() method of the RentalShop class has been revised
to ensure that the imported data uses the first column of the
CSV files as an index (as it should be). The display_stock()
method had to be adjusted accordingly.

Also the _available_cars() method was created as a "helper"
function, allowing just the available stock of cars to be obtained
rather than the entire dataset.