Changed code to uppercase abbreviated car types
One of the possible type of cars that can be rented is a
sport utility vehicle, which is abbreviated as SUV. The 'car_types'
file stores this type as 'suv'. So when displaying this type to the
user, it would need to be cast to an uppercase.

Rather than doing so just for 'SUV', a more general method has been
incorporated so that any future car types stored on the system
that are also abbreviations would be cast to uppercase. This was
accomplished by creating a new field of the 'car_types' file called
'abbrev', which stores a flag used to determine whether the name
of a car type is the abbreviated form of its true name.

Therefore the display_stock() and process_request() methods of
the RentalShop class have been slightly revised. A new auxiliary
method called _get_abbrev_types() has been created, allowing one to
obtain a list of the car type names that are abbreviated.

(One thing to mention that was added to the previous version was a
method called _get_file_dir() which constructs the relative path
to a given database file.)