Revise the display_stock() functionality
Instead of having a RentalShop object outputting the avilable stock
and the pricing for each car type separately, they are now
outputted as one single table. In other words, they are columns of
the same table.

As a result, the auxiliary methods used for displaying information
have been replaced by a single one called _output_stock().
By default it outputt just the stock, but it has a parameter that
allows one to output the pricing information also.

In addition, a few of the "caption" messages accompanying displayed
information and notifications to the user has been slightly
re-adjusted, some of which are given a space before/after it to
improve readibility on the console.