Create implementation of the "return car" module
Functionality has been added that allows the user to return a car
they have rented. In particular, the issue_bill() method of the
RentalShop class has been properly implemented. The system asks for
the vehicle registration number of the car they have rented, and
then after checking that it is recorded on the database with their
customer number, it removes the record from the system and then
outputs a bill on the console.

There was a problem concerning pd.read_csv(), as it would
automatically convert some of the fields into an undesired type.
This was most critical for the "customer_number" field of car
rentals, which should consist of strings. But they were
automatically converted to integers (as they are six-digit
numbers), which caused problems when comparing a value from it
(int) with a customer's number (str). A quick fix was made to this
problem in the _get_data() method.

A helper method called "_store_data()" has been create as an easier
way to update the contents of one of the flat files.