Add shop_db.py for dedicated database management
The rental_shop.py module (and the RentalShop class) became very
complicated, because not only was it acting as an interface for
the Customer class to access the rental shop services, but also
managed the entire database storing car data and rentals.

Therefore a new module called shop_db.py was created, containing
a class called ShopDatabase dedicated for managing the entire
database, while acting as an interface for a RentalShop object
to access its flat files and update rental data.

Therefore most the code from rental_shop.py has been ported to
this new module, including:

Methods:
- __check_db_exists()
- _get_data()
- _store_data()
- _get_file_dir()

Module Functions:
- _setup_cars()
- _setup_rentals()
- _setup_car_types()
- _random_car_ids()

Constants (all):
- _DATABASE_DIRECTORY
- _CARS
- _CAR_RENTALS
- _CAR_TYPES

The _abbev_types() method has been kept in RentalShop, as it had
a very speciifc purpose not related to retrieving one of the
flat files.

The ShopDatabase class has the following "interface" for a
RentalShop object:

- get_cars()
- get_car_rentals()
- get_car_types()
- get_available_cars()
- update_rentals()

Finally the RentalShop class has a new field representing its
database contining the cars and rentals data. The majority of its
methods had to be slightly adapted to accomodate these new changes
(in particular, accessing the database simply by using the methods
provided by this database field).