Update shop_db.py to adapt to its purpose
A number of changes have been made to the shop_db.py module
(and the ShopDatabase class), none of which heavily affect
its functionality:

* A new field called __db_dir was added to ShopDatabase, storing
the relative path to the directory where the the database's files
will be stored.

* The thee database setup functions were changed to be methods of
the ShopDatabase class. They no longer require an argument to
the directory of the data file, as they can just use the
_store_data() method to create the data files.

* A new constant called DTYPES was created. It is a dictionary
containing the data types of each column for each database file.
It is used in the _get_data() method as it reads the CSV files,
to ensure that the imported data is of the right type (and thus
avoid the previously mentioned problem with the "customer_number"
field of "car_rentals".csv).