Add VIP customer functionality
The program has been modified to allow the user to enter their
"loyalty card", which enables them to receive discounts from
car rentals (in particular, they will receive the "VIP Rate").

To represent such customers, a new subclass of the Customer class
has been created, called VIPCustomer. It differs as follows:

* It has a new field called card_number, representing the loyalty
card.

* The rent_car() method has been overridden, now passing
card_number as an additional argument to the
RentalShop.process_request() method (it has been given a new
parameter representing the customer's loyalty card number if they
have one, which gives them access to the VIP rates.)

To be honest, this does not really justify creating a new class
(the field could have simply been given to the Customer base class
with a default value as None), but nevertheless it may become
necessary to have this class if the applicaion is further
developed.