Fixed two bugs whe returning car and vaditating
Two bugs found have been fixed:

1. When a car is returned, the record is not actually removed from
the database. This is because the pandas.DataFrame drop() method
is not done in-place by default, but outputs a copy with the
applied changes.

2. The 10-digit requirement for loyalty card numbers was not put
in-place. Due to the wrong comparison operator being applied,
instead it was only inputs that were NOT 10 digits long that were
accepted.